chore: manage with memberIn memberOut

+++ Registration/src/Service/Registration.Handlers/Handlers/Registrations/MemberOutHandler.cs
@@ -1,19 +1,20 @@
 ﻿using Registration.DomainBase.Entities.Registrations;
 using Registration.DomainCore.ContextAbstraction;
 using Registration.Mapper.DTOs.Registration.MemberIn;
+using Registration.Mapper.DTOs.Registration.MemberOut;
 
 namespace Registration.Handlers.Handlers.Registrations;
 
-public class MemberInHandler
+public class MemberOutHandler
 {
-    private readonly IMemberInRepository _context;
+    private readonly IMemberOutRepository _context;
 
-    public MemberInHandler(IMemberInRepository context) 
+    public MemberOutHandler(IMemberOutRepository context) 
     {
         _context = context;
     }
 
-    public async Task<bool> CreateAsync(EditMemberInDto dto)
+    public async Task<bool> CreateAsync(EditMemberOutDto dto)
     {
         dto.Validate();
         if (!dto.IsValid)
@@ -23,8 +24,8 @@ public class MemberInHandler
 
         try
         {
-            var memberIn = new MemberIn(dto.ChurchName, dto.LastPost, dto.LetterReceiver, dto.MemberId);
-            await _context.Post(memberIn)!;
+            var memberOut = new MemberOut(dto.Reason, dto.Day, dto.MemberId);
+            await _context.Post(memberOut)!;
 
             return true;
         }
@@ -34,7 +35,7 @@ public class MemberInHandler
         }
     }
 
-    public async Task<bool> Update(EditMemberInDto dto)
+    public async Task<bool> Update(EditMemberOutDto dto)
     {
         await DeleteByMemberAsync(dto.MemberId);
         await CreateAsync(dto);
@@ -45,12 +46,14 @@ public class MemberInHandler
     public async Task<bool> DeleteByMemberAsync(int memberId)
     {
         var memberIn = await GetOneByMemberAsync(memberId);
+        if(memberIn == null) 
+            return true; 
 
         await _context.Delete(memberIn);
         return true;
     }
 
-    public async Task<MemberIn> GetOneByMemberAsync(int memberId)
+    public async Task<MemberOut> GetOneByMemberAsync(int memberId)
     {
         var memberIn = await _context.GetOneByMemberAsync(memberId);
         return memberIn!;
